в надежде пройти последний барьер

+++ kaseev.andrey/S1/AdditionalFunctions.hpp
@@ -4,12 +4,12 @@
 
 namespace kaseev
 {
-  int readList(const std::string &line, kaseev::List<std::pair<std::string, kaseev::List<int>>>& arr, bool &marker);
+  int readList(const std::string &line, kaseev::List<std::pair<std::string, kaseev::List<unsigned long long>>>& arr, bool &marker);
   int sumNumbersInString(const std::string& line);
   void sumNumbersInArray(const kaseev::List<std::pair<std::string, kaseev::List<int>>>& sums, bool &marker);
-  void printListNames(const kaseev::List<std::pair<std::string, kaseev::List<int>>>& arr);
+  void printListNames(const kaseev::List<std::pair<std::string, kaseev::List<unsigned long long>>>& arr);
   kaseev::List<std::pair<std::string, kaseev::List<int>>>
-  calculateSumList(const kaseev::List<std::pair<std::string, kaseev::List<int>>>& arr);
+  calculateSumList(const kaseev::List<std::pair<std::string, kaseev::List<unsigned long long>>>& arr);
 }
 
 #endif
